+++ Web/App/entities/ElementField.js
@@ -176,7 +176,7 @@
                     : 50; // Default value?
             }
 
-            // TODO Since this is a fixed value based on IndexRating & current user's rate,
+            // TODO Since this is a fixed value based on IndexRatingTotal & current user's rate,
             // it could be calculated on server, check it later again / SH - 03 Aug. '15
             self.otherUsersIndexRatingTotal = function () {
 
@@ -190,7 +190,7 @@
 
             self.setOtherUsersIndexRatingTotal = function () {
 
-                self.backingFields._otherUsersIndexRatingTotal = self.IndexRating;
+                self.backingFields._otherUsersIndexRatingTotal = self.IndexRatingTotal;
 
                 // Exclude current user's
                 if (self.userElementField() !== null) {